Refactoring and renaming
